And now properly test the new option *and* geneve option stacks
